third_party/WebKit/LayoutTests/typedcssom/the-stylepropertymap/computed-expected.txt

This is a testharness.js-based test.
FAIL Computed StylePropertyMap contains every CSS property assert_equals: expected 297 but got 295
PASS Computed StylePropertyMap contains CSS property declarations in style rules
FAIL Computed StylePropertyMap contains custom property declarations in style rules Illegal constructor
PASS Computed StylePropertyMap contains CSS property declarations in inline styles
FAIL Computed StylePropertyMap contains custom property declarations in inline rules Illegal constructor
PASS Computed StylePropertyMap contains computed values and not resolved values
PASS Computed StylePropertyMap is live
Harness: the test ran to completion.
